Treesame-Commit-Id: 20697d5d1ca6b4a8c7664d39c9ace4f093e71935

+++ lib/api/v1/assignment_override.rb
@@ -127,10 +127,10 @@ module Api::V1::AssignmentOverride
         errors << "invalid student_ids #{student_ids.inspect}"
         students = []
       else
-        # look up all the active students since the assignment will affect all
-        # active students in the course on this override and not just what the
+        # look up all students since the assignment will affect all current and
+        # previous students in the course on this override and not just what the
         # teacher can see that were sent in the request object
-        students = api_find_all(assignment.context.students.active, student_ids)
+        students = api_find_all(assignment.context.all_students, student_ids)
         students = students.distinct if students.is_a?(ActiveRecord::Relation)
         students = students.uniq if students.is_a?(Array)
 
@@ -427,6 +427,8 @@ module Api::V1::AssignmentOverride
       assignment.errors.add(:base, error)
     end
 
+    raise ActiveRecord::RecordInvalid.new(assignment) if assignment.errors.any?
+
     if prepared_overrides[:overrides_to_delete].any?
       assignment.assignment_overrides.where(id: prepared_overrides[:overrides_to_delete]).destroy_all
     end
@@ -436,6 +438,7 @@ module Api::V1::AssignmentOverride
     prepared_overrides[:overrides_to_create].each(&:save!)
     prepared_overrides[:overrides_to_update].each(&:save!)
 
+    assignment.touch # invalidate cached list of overrides for the assignment
     assignment.run_if_overrides_changed_later!
   end
